requirements.txt
chdp==4.0.5
asyncio==3.4.3
discord.py==1.3.4